third_party/WebKit/LayoutTests/platform/mac-mac10.11/virtual/outofblink-cors-http/http/tests/htmlimports/import-script-block-crossorigin-dynamic-expected.txt

This test ensures that crossorigin-marked script elements are blocked properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.externalScriptHasRun is undefined
PASS document.corsExternalScriptHasRun is undefined
PASS document.corsExternalScriptForLocalhostHasRun is true
PASS successfullyParsed is true

TEST COMPLETE
